Split data processing / checks to separate module
